requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
filelock==3.9.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.2.1
networkx==3.0
numpy==1.25.0
Pillow==9.5.0
requests==2.28.1
sympy==1.11.1
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.3.6
